requirements.txt
streamlit
pypdfium2
pillow
numpy